backend/requirements.txt
flask>=3.0.0
beautifulsoup4>=4.9.0
PyPDF2>=3.0.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.5
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
